Add files via upload
This simulation models a gas of particles with finite radius and no intermolecular attraction. It is a model of the Van der Waals equation with *a* set equal to zero. The .py file contains the functions needed to model the kinetic gas and must be imported and implemented to run.